+++ CHANGELOG.md
@@ -15,6 +15,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 * Adjusted input padding for better icon alignment.
 * Fixed JS function for date/time inputs so it works correctly on mobile.
 * Fixed breaking-behavior for long table-headers on mobile devices
+* Add element to upload multiple files within forms
 
 
 ## [1.4.15] - 09.10.2025